+++ pkg/manifests/eksd/images.go
