+++ pkg/manifests/eksd/read.go
